+++ cmd/gogoodwe/utils/args.go
@@ -27,7 +27,8 @@ package utils
 var (
 	// Version string
 	VersionString string = "v0.0.1"
-	infoString    string = "*GoGoodwe - A command line tool to query your SEMS Solar inverter.*"
+	infoString    string = "GoGoodwe - A CLI tool to query your SEMS Solar Inverter API."
+	reportString  string = "Report Types: (0)-Detail, (1)-Summary, (2)-Point, (3)-Plant, (4)-PlantChart, (5)-PowerFlow"
 )
 
 // Args - struct using go-arg- https://github.com/alexflint/go-arg
@@ -35,7 +36,7 @@ type Args struct {
 	Account        string `arg:"required,-a,--account" help:"SEMS Email Account."`
 	Password       string `arg:"required,-p,--password" help:"SEMS Account password."`
 	PowerStationID string `arg:"required,-i,--powerstationid" help:"SEMS Powerstation ID."`
-	DailySummary   bool   `arg:"-s,--summary" help:"Output as a daily summary."`
+	ReportType     int    `arg:",-r,--reporttype" help:"Inverter Report Number"`
 }
 
 // Description returns a command line tool to query the GOODWE SEMS Portal APIs and Solar SEMS API.
@@ -43,7 +44,7 @@ type Args struct {
 // No parameters.
 // Returns a string.
 func (Args) Description() string {
-	return infoString
+	return infoString + "\n" + reportString
 }
 
 // Version returns the version string of the Args struct.