+++ pkg/interfaces/semsdata.go
@@ -24,10 +24,15 @@ SOFTWARE.
 package interfaces
 
 import (
-	"github.com/AaronSaikovski/gogoodwe/internal/pkg/types"
+	inverteallpoint "github.com/AaronSaikovski/gogoodwe/pkg/inverterallpoint"
+	"github.com/AaronSaikovski/gogoodwe/pkg/monitordetail"
+	"github.com/AaronSaikovski/gogoodwe/pkg/monitorsummary"
+	"github.com/AaronSaikovski/gogoodwe/pkg/plantdetail"
+	plantchartdata "github.com/AaronSaikovski/gogoodwe/pkg/plantpowerchart"
+	"github.com/AaronSaikovski/gogoodwe/pkg/powerflow"
 )
 
-// Constraints for functions that return data from the API via marshalled structs
+// SEMS Data Constraints
 type SemsDataConstraint interface {
-	types.InverterData | types.DailySummaryData
+	inverteallpoint.InverterAllPoint | monitordetail.MonitorData | monitorsummary.DailySummaryData | plantdetail.PlantDetailByPowerstationId | plantchartdata.PlantPowerChart | powerflow.Powerflow
 }