+++ pkg/inverterallpoint/getinverterallpointdata.go
@@ -1,0 +1,68 @@
+/*
+MIT License
+
+# Copyright (c) 2024 Aaron Saikovski
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+*/
+
+package inverteallpoint
+
+import (
+	"github.com/AaronSaikovski/gogoodwe/pkg/apihelpers"
+	"github.com/AaronSaikovski/gogoodwe/pkg/auth"
+)
+
+const (
+
+	// Powerstation API Url
+	powerStationURL string = "/v3/PowerStation/GetInverterAllPoint"
+
+	// Default timeout value
+	HTTPTimeout int = 20
+)
+
+// GetMonitorData retrieves monitor data using login credentials and response, storing it in inverterOutput.
+//
+// Parameters:
+// - authLoginInfo: pointer to the LoginInfo struct containing the login credentials and API response
+// - inverterOutput: pointer to the data output
+// Return type: error
+func (inverterData *InverterAllPoint) GetMonitorData(authLoginInfo *auth.LoginInfo, inverterOutput interface{}) error {
+
+	return apihelpers.FetchMonitorAPIData(authLoginInfo, powerStationURL, HTTPTimeout, inverterOutput)
+}
+
+// GetPowerData retrieves the power data for a daily summary using the provided authentication information.
+//
+// Parameters:
+// - authLoginInfo: a pointer to the authentication information for the user.
+//
+// Returns:
+// - error: an error if there was a problem retrieving the power data.
+func (inverterData *InverterAllPoint) GetPowerData(authLoginInfo *auth.LoginInfo) error {
+
+	// Get monitor data
+	if err := inverterData.GetMonitorData(authLoginInfo, inverterData); err != nil {
+		return err
+	}
+
+	return apihelpers.ProcesData(inverterData)
+
+}