+++ pkg/utils/errorhandler.go
@@ -21,11 +21,19 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-package apilogin
 
-// ApiLoginCredentials - Struct to hold User login credentials
-type ApiLoginCredentials struct {
-	Account        string `json:"account"`
-	Password       string `json:"pwd"`
-	PowerStationID string `json:"powerstationid"`
+package utils
+
+import (
+	"log"
+
+	"github.com/logrusorgru/aurora"
+)
+
+// HandleError - Generic error handler.
+//
+// Parameters:
+// - err: the error to handle.
+func HandleError(err error) {
+	log.Fatal(aurora.BrightRed(err.Error()))
 }